Revert "SALTO-5761: Upgrade non guide types to be fetchable using the new infra (#5928)"
This reverts commit 69e0cba7e6159c74c7c51d3b9e188a32151e481e.
