+++ packages/zendesk-adapter/src/change_validator.ts
@@ -88,8 +88,7 @@ import {
   viewCustomStatusConditionsValidator,
 } from './change_validators'
 import ZendeskClient from './client/client'
-import { ChangeValidatorName, ZendeskConfig } from './config'
-import { ZendeskDeployConfig, ZendeskFetchConfig } from './user_config'
+import { ChangeValidatorName, ZendeskDeployConfig, ZendeskFetchConfig, ZendeskConfig } from './config'
 
 const {
   deployTypesNotSupportedValidator,
