+++ packages/zendesk-adapter/src/change_validators/child_parent/child_missing_parent_annotation.ts
@@ -30,7 +30,7 @@ import {
   isAdditionChange,
 } from '@salto-io/adapter-api'
 import { getParent } from '@salto-io/adapter-utils'
-import { ZendeskApiConfig } from '../../user_config'
+import { ZendeskApiConfig } from '../../config'
 import { getChildAndParentTypeNames } from './utils'
 
 const createChildReferencesError = (
