+++ packages/zendesk-adapter/src/change_validators/child_parent/missing_from_parent.ts
@@ -27,7 +27,7 @@ import {
   isAdditionChange,
   isAdditionOrModificationChange,
 } from '@salto-io/adapter-api'
-import { ZendeskApiConfig } from '../../user_config'
+import { ZendeskApiConfig } from '../../config'
 import { getChildAndParentTypeNames, getRemovedAndAddedChildren } from './utils'
 
 export const createParentReferencesError = (
