+++ packages/zendesk-adapter/src/change_validators/child_parent/parent_annotation_has_single_value.ts
@@ -21,7 +21,7 @@ import {
   CORE_ANNOTATIONS,
   isReferenceExpression,
 } from '@salto-io/adapter-api'
-import { ZendeskApiConfig } from '../../user_config'
+import { ZendeskApiConfig } from '../../config'
 import { getChildAndParentTypeNames } from './utils'
 
 export const parentAnnotationToHaveSingleValueValidatorCreator =
