+++ packages/zendesk-adapter/src/change_validators/child_parent/removed_from_parent.ts
@@ -22,7 +22,7 @@ import {
   isRemovalChange,
   ChangeError,
 } from '@salto-io/adapter-api'
-import { ZendeskApiConfig } from '../../user_config'
+import { ZendeskApiConfig } from '../../config'
 import { getChildAndParentTypeNames, getRemovedAndAddedChildren, ChildParentRelationship } from './utils'
 import { FIELD_TYPE_NAMES } from '../../constants'
