+++ packages/zendesk-adapter/src/change_validators/custom_role_removal.ts
@@ -20,7 +20,7 @@ import { getUsers } from '../user_utils'
 import { paginate } from '../client/pagination'
 import ZendeskClient from '../client/client'
 import { CUSTOM_ROLE_TYPE_NAME } from '../constants'
-import { ZendeskFetchConfig } from '../user_config'
+import { ZendeskFetchConfig } from '../config'
 
 const { createPaginator } = clientUtils
