+++ packages/zendesk-adapter/src/change_validators/duplicate_id_field_values.ts
@@ -27,7 +27,7 @@ import { elements as elementUtils } from '@salto-io/adapter-components'
 import { logger } from '@salto-io/logging'
 import { values as lowerDashValues } from '@salto-io/lowerdash'
 import { getInstancesFromElementSource } from '@salto-io/adapter-utils'
-import { ZendeskApiConfig } from '../user_config'
+import { ZendeskApiConfig } from '../config'
 
 const { generateInstanceNameFromConfig } = elementUtils
 const { isDefined } = lowerDashValues
