+++ packages/zendesk-adapter/src/change_validators/guide_creation_or_removal.ts
@@ -25,7 +25,7 @@ import {
 } from '@salto-io/adapter-api'
 import { BRAND_TYPE_NAME } from '../constants'
 import ZendeskClient from '../client/client'
-import { ZendeskApiConfig } from '../user_config'
+import { ZendeskApiConfig } from '../config'
 import { isBrand, invalidBrandChange } from './guide_activation'
 
 export const invalidBrandAdditionChange = (
