+++ packages/zendesk-adapter/src/change_validators/target.ts
@@ -22,7 +22,7 @@ import {
   isAdditionOrModificationChange,
   isInstanceChange,
 } from '@salto-io/adapter-api'
-import { ZendeskApiConfig } from '../user_config'
+import { ZendeskApiConfig } from '../config'
 import ZendeskClient from '../client/client'
 import { TARGET_TYPE_NAME } from '../constants'
