+++ packages/zendesk-adapter/src/change_validators/trigger_category_removal.ts
@@ -27,9 +27,8 @@ import { logger } from '@salto-io/logging'
 import { values as lowerDashValues } from '@salto-io/lowerdash'
 import { definitions } from '@salto-io/adapter-components'
 import { getInstancesFromElementSource } from '@salto-io/adapter-utils'
-import { OMIT_INACTIVE_DEFAULT } from '../config'
 import { TRIGGER_CATEGORY_TYPE_NAME, TRIGGER_TYPE_NAME } from '../constants'
-import { ZendeskFetchConfig } from '../user_config'
+import { OMIT_INACTIVE_DEFAULT, ZendeskFetchConfig } from '../config'
 
 const { isDefined } = lowerDashValues
 const log = logger(module)
