+++ packages/zendesk-adapter/src/deployment.ts
@@ -31,7 +31,7 @@ import { logger } from '@salto-io/logging'
 import { collections } from '@salto-io/lowerdash'
 import ZendeskClient from './client/client'
 import { getZendeskError } from './errors'
-import { ZendeskApiConfig } from './user_config'
+import { ZendeskApiConfig } from './config'
 
 const log = logger(module)
 const { awu } = collections.asynciterable
