+++ packages/zendesk-adapter/src/filters/custom_objects/custom_object_fields_order.ts
@@ -34,7 +34,6 @@ import { getParent, getParents, inspectValue, isResolvedReferenceExpression } fr
 import { values as lowerDashValues } from '@salto-io/lowerdash'
 import { elements as elementsUtils } from '@salto-io/adapter-components'
 import { logger } from '@salto-io/logging'
-import { FETCH_CONFIG } from '../../config'
 import { FilterCreator } from '../../filter'
 import {
   CUSTOM_OBJECT_FIELD_ORDER_TYPE_NAME,
@@ -61,7 +60,7 @@ export const customObjectFieldsOrderType = new ObjectType({
  The fields are returned from the api by order, so we save it to be able to properly reorder them on deploy
  This is needed because two fields can have the same positions, and then be sorted by non multi-env fields
  */
-const customObjectFieldsOrderFilter: FilterCreator = ({ client, config }) => ({
+const customObjectFieldsOrderFilter: FilterCreator = ({ client }) => ({
   name: 'customObjectFieldsOrderFilter',
   onFetch: async (elements: Element[]) => {
     const customObjectFields = elements
@@ -80,7 +79,7 @@ const customObjectFieldsOrderFilter: FilterCreator = ({ client, config }) => ({
     if (!_.isEmpty(customObjectFieldsByParentName)) {
       elements.push(customObjectFieldsOrderType)
     }
-    const usingNewInfra = config[FETCH_CONFIG].useNewInfra
+
     Object.entries(customObjectFieldsByParentName).forEach(([parentName, fields]) => {
       const parent = customObjectsByFullName[parentName]
       if (parent === undefined) {
@@ -92,10 +91,7 @@ const customObjectFieldsOrderFilter: FilterCreator = ({ client, config }) => ({
         instanceName,
         customObjectFieldsOrderType,
         {
-          // in the new infra, the order of the fields is reversed in the element map
-          [ORDER_FIELD]: (usingNewInfra === true ? fields.reverse() : fields).map(
-            field => new ReferenceExpression(field.elemID, field),
-          ),
+          [ORDER_FIELD]: fields.map(field => new ReferenceExpression(field.elemID, field)),
         },
         [ZENDESK, RECORDS_PATH, CUSTOM_OBJECT_FIELD_ORDER_TYPE_NAME, instanceName],
         {
