+++ packages/zendesk-adapter/test/change_validators/trigger_category_removal.test.ts
@@ -18,8 +18,7 @@ import { elementSource as elementSourceUtils } from '@salto-io/workspace'
 import _ from 'lodash'
 import { TRIGGER_CATEGORY_TYPE_NAME, TRIGGER_TYPE_NAME, ZENDESK } from '../../src/constants'
 import { triggerCategoryRemovalValidator } from '../../src/change_validators'
-import { DEFAULT_CONFIG, FETCH_CONFIG } from '../../src/config'
-import { ZendeskFetchConfig } from '../../src/user_config'
+import { DEFAULT_CONFIG, FETCH_CONFIG, ZendeskFetchConfig } from '../../src/config'
 
 const { createInMemoryElementSource } = elementSourceUtils
