+++ packages/zendesk-adapter/test/change_validators/users.test.ts
@@ -18,7 +18,7 @@ import { elementSource } from '@salto-io/workspace'
 import { ZENDESK, ARTICLE_TYPE_NAME } from '../../src/constants'
 import { usersValidator } from '../../src/change_validators'
 import ZendeskClient from '../../src/client/client'
-import { ZendeskFetchConfig } from '../../src/user_config'
+import { ZendeskFetchConfig } from '../../src/config'
 
 const { createInMemoryElementSource } = elementSource
