+++ packages/zendesk-adapter/test/inactive.test.ts
@@ -15,10 +15,9 @@
  */
 import { ElemID, InstanceElement, ObjectType } from '@salto-io/adapter-api'
 import _ from 'lodash'
-import { ValueGeneratedItem } from '@salto-io/adapter-components/src/fetch'
 import { DEFAULT_CONFIG, FETCH_CONFIG } from '../src/config'
 import { ZENDESK } from '../src/constants'
-import { filterOutInactiveInstancesForType, filterOutInactiveItemForType } from '../src/inactive'
+import { filterOutInactiveInstancesForType } from '../src/inactive'
 
 describe('omit inactive', () => {
   const trigger = new ObjectType({ elemID: new ElemID(ZENDESK, 'trigger') })
@@ -36,121 +35,14 @@ describe('omit inactive', () => {
   const ticketForm = new ObjectType({ elemID: new ElemID(ZENDESK, 'ticket_form') })
   const ticketForm1 = new InstanceElement('inst1', ticketForm, { name: 'test', active: false })
 
-  const activeTicketFieldItem = {
-    value: {
-      id: 123,
-      type: 'tagger',
-      title: 'macroTest',
-      raw_title: 'macroTest',
-      active: true,
-      required: false,
-    },
-    typeName: 'ticket_field',
-    context: {},
-  }
-
-  const activeCustomRoleItem = {
-    value: {
-      id: 123,
-      name: 'testForFailure',
-      description: 'testForFailureDescription',
-      configuration: {
-        chat_access: true,
-        end_user_list_access: 'full',
-        ticket_tag_editing: false,
-      },
-      team_member_count: 0,
-    },
-    typeName: 'custom_role',
-    context: {},
-  }
-
-  const activeOrganizationFieldItem = {
-    value: {
-      id: 123,
-      type: 'text',
-      title: 'fieldTest',
-      active: true,
-      system: false,
-    },
-    typeName: 'organization_field',
-    context: {},
-  }
-
-  const inactiveMacroItem = {
-    value: {
-      id: 123,
-      title: 'test',
-      active: false,
-      default: false,
-      description: null,
-    },
-    typeName: 'macro',
-    context: {},
-  }
-
-  const inactiveViewItem = {
-    value: {
-      id: 123,
-      title: 'Pending tickets',
-      active: false,
-      default: false,
-      description: null,
-      restriction: null,
-      raw_title: 'Pending tickets',
-    },
-    typeName: 'view',
-    context: {},
-  }
-
-  const activeWebhookItem = {
-    value: {
-      id: 'abc',
-      name: 'Updated',
-      status: 'active',
-      subscriptions: ['conditional_ticket_events'],
-      endpoint: 'https://example.com/status/200',
-      http_method: 'POST',
-      request_format: 'json',
-    },
-    typeName: 'webhook',
-    context: {},
-  }
-
-  const inactiveWebhookItem = {
-    value: {
-      id: 'abc',
-      name: 'Updated',
-      status: 'inactive',
-      subscriptions: ['conditional_ticket_events'],
-      endpoint: 'https://example.com/status/200',
-      http_method: 'POST',
-      request_format: 'json',
-    },
-    typeName: 'webhook',
-    context: {},
-  }
-
-  const valueGeneratedItems = [
-    activeTicketFieldItem,
-    activeCustomRoleItem,
-    activeOrganizationFieldItem,
-    inactiveMacroItem,
-    inactiveViewItem,
-    activeWebhookItem,
-    inactiveWebhookItem,
-  ]
-
   describe('onFetch', () => {
     let instanceFilter: (instances: InstanceElement[]) => InstanceElement[]
-    let itemFilter: (item: ValueGeneratedItem) => boolean
     beforeEach(async () => {
       jest.clearAllMocks()
     })
     describe('using default config', () => {
       beforeEach(async () => {
         instanceFilter = filterOutInactiveInstancesForType(DEFAULT_CONFIG)
-        itemFilter = filterOutInactiveItemForType(DEFAULT_CONFIG)
       })
       it('should omit inactive instances if omitInactive in typeDefaults is true', async () => {
         expect(instanceFilter([trigger1, trigger2]).map(elem => elem.elemID.getFullName())).toEqual([
@@ -165,12 +57,6 @@ describe('omit inactive', () => {
           webhook3.elemID.getFullName(),
         ])
         expect(instanceFilter([view1]).map(elem => elem.elemID.getFullName())).toEqual([])
-        expect(valueGeneratedItems.filter(item => itemFilter(item))).toEqual([
-          activeTicketFieldItem,
-          activeCustomRoleItem,
-          activeOrganizationFieldItem,
-          activeWebhookItem,
-        ])
       })
       it('should not omit instance of types that we need their inactive instances for reorder', async () => {
         expect(instanceFilter([ticketForm1]).map(elem => elem.elemID.getFullName())).toEqual([
@@ -187,7 +73,6 @@ describe('omit inactive', () => {
       it('should not omit instance if it does not have active field', async () => {
         const inst = new InstanceElement('inst1', trigger, { name: 'test' })
         expect(instanceFilter([inst]).map(elem => elem.elemID.getFullName())).toEqual([inst.elemID.getFullName()])
-        expect(itemFilter(activeOrganizationFieldItem)).toBeTruthy()
       })
     })
     describe('using custom config', () => {
@@ -196,9 +81,8 @@ describe('omit inactive', () => {
           const config = _.cloneDeep(DEFAULT_CONFIG)
           config[FETCH_CONFIG].omitInactive = {
             default: false,
-            customizations: { trigger: true, macro: true, webhook: true, view: true },
+            customizations: { trigger: true, macro: true, webhook: true },
           }
-          itemFilter = filterOutInactiveItemForType(config)
           instanceFilter = filterOutInactiveInstancesForType(config)
         })
         it('should omit inactive instances if their customizations is true', async () => {
@@ -212,12 +96,6 @@ describe('omit inactive', () => {
             webhook1.elemID.getFullName(),
             webhook3.elemID.getFullName(),
           ])
-          expect(valueGeneratedItems.filter(item => itemFilter(item))).toEqual([
-            activeTicketFieldItem,
-            activeCustomRoleItem,
-            activeOrganizationFieldItem,
-            activeWebhookItem,
-          ])
         })
       })
       describe('with omitInactive set to false', () => {
@@ -226,7 +104,6 @@ describe('omit inactive', () => {
           config[FETCH_CONFIG].omitInactive = {
             customizations: { trigger: false, macro: false, webhook: false },
           }
-          itemFilter = filterOutInactiveItemForType(config)
           instanceFilter = filterOutInactiveInstancesForType(config)
         })
         it('should omit inactive instances if their customizations is true', async () => {
@@ -243,7 +120,6 @@ describe('omit inactive', () => {
             webhook2.elemID.getFullName(),
             webhook3.elemID.getFullName(),
           ])
-          expect(valueGeneratedItems.filter(item => itemFilter(item))).toEqual(valueGeneratedItems)
         })
       })
     })
